# codehelp-admin-bs/codehelp-admin-api-docs/api/code-help-core/api/openapi.yaml
openapi: 3.0.1
info:
  title: Coding helper spec
  version: 1.0.0
servers:
- description: Open API specification
  url: http://localhost:3000/api
security:
- Authorization_JWT: []
paths:
...
  /contests:
    get:
      description: Returns the contests happening in coding helper
      operationId: getContestEntries
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestsResponse'
          description: A list of contests
      tags:
      - Contest
    post:
      description: Creates a contest
      operationId: createContestEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContestBaseEntry'
      responses:
        "204":
          description: No content
      tags:
      - Contest
  /contests/{id}:
    delete:
      description: Deletes contest happening in coding helper
      operationId: deleteContestEntry
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "204":
          description: No content
      tags:
      - Contest
    get:
      description: Returns contest happening in coding helper
      operationId: getContestEntry
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestEntry'
          description: A list of contests
      tags:
      - Contest
    post:
      description: Creates contest problem entry for a contest
      operationId: createContestProblemEntry
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProblemReqBody'
      responses:
        "204":
          description: No content
      tags:
      - Contest
    put:
      description: Modifies contest happening in coding helper
      operationId: editContestEntry
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContestBaseEntry'
      responses:
        "204":
          description: No content
      tags:
      - Contest
  /contest/{id}/start:
    get:
      description: Starts contest happening in coding helper
      operationId: startContest
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "204":
          description: No content
      tags:
      - Contest
  /contest/{id}/close:
    get:
      description: Closes contest happening in coding helper
      operationId: closeContest
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "204":
          description: No content
      tags:
      - Contest
  /contest/{contestId}/problems/{problemId}:
    delete:
      description: Returns a contest problem entry
      operationId: deleteContestProblem
      parameters:
      - explode: false
        in: path
        name: contestId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      - explode: false
        in: path
        name: problemId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "204":
          description: No content
      tags:
      - Contest
    get:
      description: Returns a contest problem entry
      operationId: getContestProblem
      parameters:
      - explode: false
        in: path
        name: contestId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      - explode: false
        in: path
        name: problemId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestProblem'
          description: Contest problem entry
      tags:
      - Contest
  /contest/{contestId}/problems/{problemId}/set-score:
    post:
      description: Sets the score of a contest problem entry
      operationId: setContestProblemScore
      parameters:
      - explode: false
        in: path
        name: contestId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      - explode: false
        in: path
        name: problemId
        required: true
        schema:
          $ref: '#/components/schemas/Id'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/setContestProblemScore_request'
      responses:
        "204":
          description: No content
      tags:
      - Contest
...
components:
...
  schemas:
    Id:
      example: 23
      format: long
      type: integer
    Difficulty:
      enum:
      - EASY
      - MEDIUM
      - HARD
      example: EASY
      type: string
...
    File:
      format: binary
      type: string
...
    Category:
      example:
        name: Binary trees
      properties:
        name:
          example: Binary trees
          type: string
      type: object
    ProblemBaseEntry:
      properties:
        category:
          $ref: '#/components/schemas/Category'
        title:
          example: Two sum
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        markdown:
          example: '## Two Sum... problem description'
          type: string
      required:
      - difficulty
      - title
      type: object
    ProblemEntry:
      allOf:
      - $ref: '#/components/schemas/ProblemBaseEntry'
      - properties:
          id:
            $ref: '#/components/schemas/Id'
          starterCode:
            example: "function(...) {...}"
            type: string
        required:
        - starterCode
        type: object
      example:
        difficulty: EASY
        starterCode: "function(...) {...}"
        markdown: '## Two Sum... problem description'
        id: 23
        category:
          name: Binary trees
        title: Two sum
...
    ProblemReqBody:
      allOf:
      - $ref: '#/components/schemas/ProblemBaseEntry'
      - properties:
          starterCode:
            $ref: '#/components/schemas/File'
          runnerCode:
            $ref: '#/components/schemas/File'
          testCases:
            items:
              $ref: '#/components/schemas/File'
            type: array
        required:
        - runnerCode
        - starterCode
        - testCases
        type: object
...
    ContestStatus:
      enum:
      - OPEN
      - STARTED
      - CLOSED
      type: string
    ContestProblem:
      example:
        score: 53
        problem:
          difficulty: EASY
          starterCode: "function(...) {...}"
          markdown: '## Two Sum... problem description'
          id: 23
          category:
            name: Binary trees
          title: Two sum
        id: 23
      properties:
        id:
          example: 23
          format: long
          type: integer
        problem:
          $ref: '#/components/schemas/ProblemEntry'
        score:
          example: 53
          format: long
          type: integer
      required:
      - problem
      - score
      type: object
    ContestBaseEntry:
      example:
        duration: 1h30m
        name: Bitweekly 122 contest
        startsOn: 2022-05-16
      properties:
        name:
          example: Bitweekly 122 contest
          type: string
        duration:
          example: 1h30m
          format: duration
          type: string
        startsOn:
          example: 2022-05-16
          format: date
          type: string
      required:
      - duration
      - name
      - startsOn
      type: object
    ContestEntry:
      allOf:
      - $ref: '#/components/schemas/ContestBaseEntry'
      - properties:
          id:
            $ref: '#/components/schemas/Id'
          status:
            $ref: '#/components/schemas/ContestStatus'
          problems:
            items:
              $ref: '#/components/schemas/ContestProblem'
            type: array
        required:
        - problems
        type: object
      example:
        duration: 1h30m
        name: Bitweekly 122 contest
        id: 23
        startsOn: 2022-05-16
        status: null
        problems:
        - score: 53
          problem:
            difficulty: EASY
            starterCode: "function(...) {...}"
            markdown: '## Two Sum... problem description'
            id: 23
            category:
              name: Binary trees
            title: Two sum
          id: 23
        - score: 53
          problem:
            difficulty: EASY
            starterCode: "function(...) {...}"
            markdown: '## Two Sum... problem description'
            id: 23
            category:
              name: Binary trees
            title: Two sum
          id: 23
      type: object
...
    ContestsResponse:
      example:
        contests:
        - duration: 1h30m
          name: Bitweekly 122 contest
          id: 23
          startsOn: 2022-05-16
          status: null
          problems:
          - score: 53
            problem:
              difficulty: EASY
              starterCode: "function(...) {...}"
              markdown: '## Two Sum... problem description'
              id: 23
              category:
                name: Binary trees
              title: Two sum
            id: 23
          - score: 53
            problem:
              difficulty: EASY
              starterCode: "function(...) {...}"
              markdown: '## Two Sum... problem description'
              id: 23
              category:
                name: Binary trees
              title: Two sum
            id: 23
        - duration: 1h30m
          name: Bitweekly 122 contest
          id: 23
          startsOn: 2022-05-16
          status: null
          problems:
          - score: 53
            problem:
              difficulty: EASY
              starterCode: "function(...) {...}"
              markdown: '## Two Sum... problem description'
              id: 23
              category:
                name: Binary trees
              title: Two sum
            id: 23
          - score: 53
            problem:
              difficulty: EASY
              starterCode: "function(...) {...}"
              markdown: '## Two Sum... problem description'
              id: 23
              category:
                name: Binary trees
              title: Two sum
            id: 23
      properties:
        contests:
          items:
            $ref: '#/components/schemas/ContestEntry'
          type: array
      required:
      - contests
      type: object
...
    setContestProblemScore_request:
      properties:
        score:
          example: 15
          type: integer
      type: object
...
  securitySchemes:
    Authorization_JWT:
      bearerFormat: JWT
      scheme: bearer
      type: http