# codehelp-admin-bs/codehelp-admin-api-docs/api/code-help-core/go/model_contest_base_entry.go
/*
 * Coding helper spec
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package codehelp_core_api




type ContestBaseEntry struct {

	Name string `json:"name"`

	Duration string `json:"duration"`

	StartsOn string `json:"startsOn"`
}

// AssertContestBaseEntryRequired checks if the required fields are not zero-ed
func AssertContestBaseEntryRequired(obj ContestBaseEntry) error {
	elements := map[string]interface{}{
		"name": obj.Name,
		"duration": obj.Duration,
		"startsOn": obj.StartsOn,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertContestBaseEntryConstraints checks if the values respects the defined constraints
func AssertContestBaseEntryConstraints(obj ContestBaseEntry) error {
	return nil
}