# codehelp-admin-bs/codehelp-admin-api-docs/api/code-help-core/go/model_contests_response.go
/*
 * Coding helper spec
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package codehelp_core_api




type ContestsResponse struct {

	Contests []ContestEntry `json:"contests"`
}

// AssertContestsResponseRequired checks if the required fields are not zero-ed
func AssertContestsResponseRequired(obj ContestsResponse) error {
	elements := map[string]interface{}{
		"contests": obj.Contests,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	for _, el := range obj.Contests {
		if err := AssertContestEntryRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertContestsResponseConstraints checks if the values respects the defined constraints
func AssertContestsResponseConstraints(obj ContestsResponse) error {
	return nil
}